# docs/swagger.yaml
basePath: /
definitions:
...
  editorapi.ListsParams:
    properties:
      pageNo:
        type: integer
      pageSize:
        type: integer
    type: object
...
  editorapi.langDelPara:
    properties:
      lan_code:
        type: string
    type: object
  editorapi.langEdit:
    properties:
      desc:
        additionalProperties:
          type: string
        type: object
      flag:
        items:
          type: string
        type: array
      is_hot:
        type: boolean
      is_show:
        type: boolean
      list_order:
        type: integer
      title:
        additionalProperties:
          type: string
        type: object
      word_direction:
        type: string
    type: object
  editorapi.langEditPara:
    properties:
      lan_code:
        type: string
      lang_info:
        $ref: '#/definitions/editorapi.langEdit'
        type: object
    type: object
  editorapi.langPara:
    properties:
      desc:
        additionalProperties:
          type: string
        type: object
      flag:
        items:
          type: string
        type: array
      has_del:
        type: boolean
      is_hot:
        type: boolean
      is_show:
        type: boolean
      lan_code:
        type: string
      list_order:
        type: integer
      title:
        additionalProperties:
          type: string
        type: object
      word_direction:
        type: string
    type: object
...
info:
  contact: {}
  description: 全球说编辑器后台API服务文档
  license: {}
  title: 全球说编辑器后台API服务文档
  version: 0.0.1
paths:
...
  /editor/lang/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语种信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.langPara'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 添加语种
      tags:
      - EditorLangAPI（语言种类接口）
  /editor/lang/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语种信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.langDelPara'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除语种信息
      tags:
      - EditorLangAPI（语言种类接口）
  /editor/lang/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语种信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.langEditPara'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"编辑成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑语种信息
      tags:
      - EditorLangAPI（语言种类接口）
  /editor/lang/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语种列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/editorapi.ListsParams'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 编辑器语种列表
      tags:
      - EditorLangAPI（语言种类接口）
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"